+++ portal-impl/src/com/liferay/portlet/internal/PortletContainerImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet;
+package com.liferay.portlet.internal;
 
 import aQute.bnd.annotation.ProviderType;
 
@@ -67,7 +67,22 @@ import com.liferay.portal.kernel.util.comparator.PortletConfigurationIconCompara
 import com.liferay.portal.kernel.webdav.WebDAVStorage;
 import com.liferay.portal.kernel.xml.QName;
 import com.liferay.portal.theme.PortletDisplayFactory;
-import com.liferay.portlet.internal.PortletAppUtil;
+import com.liferay.portlet.ActionRequestFactory;
+import com.liferay.portlet.ActionRequestImpl;
+import com.liferay.portlet.ActionResponseFactory;
+import com.liferay.portlet.ActionResponseImpl;
+import com.liferay.portlet.EventImpl;
+import com.liferay.portlet.EventRequestFactory;
+import com.liferay.portlet.EventRequestImpl;
+import com.liferay.portlet.EventResponseFactory;
+import com.liferay.portlet.EventResponseImpl;
+import com.liferay.portlet.InvokerPortletImpl;
+import com.liferay.portlet.PublicRenderParametersPool;
+import com.liferay.portlet.RenderParametersPool;
+import com.liferay.portlet.ResourceRequestFactory;
+import com.liferay.portlet.ResourceRequestImpl;
+import com.liferay.portlet.ResourceResponseFactory;
+import com.liferay.portlet.ResourceResponseImpl;
 import com.liferay.util.SerializableUtil;
 
 import java.io.Serializable;
